% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pumsr.R
\name{pumsr}
\alias{pumsr}
\title{Import fixed-width files from ipums.org directly into R using the corresponding XML IPUMS DII codebook.}
\usage{
pumsr(dat, codebook, large = FALSE, labels = FALSE)
}
\arguments{
\item{dat}{The fixed-width data file.}

\item{codebook}{The XML version of the codebook for the corresponding ipums extract.}

\item{large}{Files designated as "large" are imported using the LaF and ffbase packages to avoid memory problems. Defaults to FALSE. Factor labels are currently not supported with large files.}

\item{labels}{If TRUE, the labels option converts character variables to factors and attaches value labels. Defaults to FALSE.}
}
\description{
Import fixed-width files from ipums.org directly into R using the corresponding XML IPUMS DII codebook.
}
\examples{
ihis <- pumsr("ihis_00001.dat", "ihis_00001.xml", labels=TRUE)
}
